implemented main coordinator
