+++ Navigation/Controllers/TabBarController.swift
@@ -10,26 +10,4 @@ import UIKit
 
 class TabBarController: UITabBarController {
 
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        let model = MyModel()
-        
-        let feedController = FeedViewController(model: model)
-        let navigationFirst = UINavigationController(rootViewController: feedController)
-        navigationFirst.tabBarItem = UITabBarItem(
-            title: "Feed",
-            image: UIImage(systemName: "house.fill"),
-            selectedImage: nil)
-        
-        let loginController = LogInViewController()
-        let navigationSecond = UINavigationController(rootViewController: loginController)
-        navigationSecond.tabBarItem = UITabBarItem(
-            title: "Profile",
-            image: UIImage(systemName: "person.fill"),
-            selectedImage: nil)
-
-        viewControllers = [navigationFirst, navigationSecond]
-    }
 }
