+++ Navigation/SceneDelegate.swift
@@ -13,7 +13,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
     private let loginFactory = MyLoginFactory()
     private let myModel = MyModel()
-
+    private let mainCoordinator = MainCoordinator()
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
@@ -21,8 +22,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let scene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: scene)
         
-        let tabBarController = TabBarController()
-        window?.rootViewController = tabBarController
+        
+        window?.rootViewController = mainCoordinator.tabBarController
         
         window?.makeKeyAndVisible()
